Finish the Button
- Rename the file, function and export to OpenPostSettingsModal. This name reflects the function of this component better without becoming too long.
- Add i18n to the button text.
- Make the button text depend on the type of the post.
- Use the correct action for updating the modal.
  - We have an eslint rule that prevents binding in the onClick. Therefore, I have created a seperate function that acts as a wrapper.
